twilio/twilio_gemini/requirements.txt
click==8.1.7
Flask==3.0.3
google-generativeai==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
six==1.16.0
twilio==9.0.0
Werkzeug==3.0.3